Refactor entry import functionality and remove unused imports
