+++ src/main/java/dev/ccosta/aisha/web/entry/EntryController.java
@@ -9,15 +9,11 @@ import dev.ccosta.aisha.application.entry.EntryCategorySelection;
 import dev.ccosta.aisha.application.entry.categorization.EntryCategorySuggestion;
 import dev.ccosta.aisha.application.entry.categorization.EntryCategorySuggestionRequest;
 import dev.ccosta.aisha.application.entry.categorization.EntryCategorySuggestionService;
-import dev.ccosta.aisha.application.entry.importing.EntryCsvImportOptions;
-import dev.ccosta.aisha.application.entry.importing.EntryImportFailureCause;
 import dev.ccosta.aisha.application.entry.transfer.EntryTransferCounterpartRequest;
 import dev.ccosta.aisha.application.entry.transfer.EntryTransferCreationRequest;
 import dev.ccosta.aisha.application.entry.transfer.EntryTransferService;
 import dev.ccosta.aisha.application.entry.transfer.EntryTransferView;
 import dev.ccosta.aisha.application.entry.EntrySettlementAfterAccountDeactivationException;
-import dev.ccosta.aisha.application.entry.importing.statement.EntryStatementFormat;
-import dev.ccosta.aisha.application.entry.importing.statement.EntryStatementImportService;
 import dev.ccosta.aisha.application.entry.EntryNotFoundException;
 import dev.ccosta.aisha.application.entry.EntryService;
 import dev.ccosta.aisha.domain.account.Account;
@@ -52,7 +48,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.multipart.MultipartFile;
 
 @Controller
 @RequestMapping("/entries")
@@ -68,8 +63,6 @@ public class EntryController {
     private final CategoryService categoryService;
     private final EntryCategorySuggestionService entryCategorySuggestionService;
     private final EntryTransferService entryTransferService;
-    private final EntryStatementImportService entryStatementImportService;
-    private final EntryImportJobCoordinator entryImportJobCoordinator;
     private final MessageSource messageSource;
 
     public EntryController(
@@ -78,8 +71,6 @@ public class EntryController {
         CategoryService categoryService,
         EntryCategorySuggestionService entryCategorySuggestionService,
         EntryTransferService entryTransferService,
-        EntryStatementImportService entryStatementImportService,
-        EntryImportJobCoordinator entryImportJobCoordinator,
         MessageSource messageSource
     ) {
         this.entryService = entryService;
@@ -87,8 +78,6 @@ public class EntryController {
         this.categoryService = categoryService;
         this.entryCategorySuggestionService = entryCategorySuggestionService;
         this.entryTransferService = entryTransferService;
-        this.entryStatementImportService = entryStatementImportService;
-        this.entryImportJobCoordinator = entryImportJobCoordinator;
         this.messageSource = messageSource;
     }
 
@@ -123,106 +112,6 @@ public class EntryController {
         fillListing(model, globalDateFilter, accountId, categoryId, description, pendingSuggestions, page, size);
         setCanonicalEntriesPushUrl(request, response);
         return "entries/list :: listing";
-    }
-
-    @GetMapping("/import")
-    public String importPage(Model model) {
-        model.addAttribute("mode", "idle");
-        return "entries/import";
-    }
-
-    @GetMapping("/statement-import")
-    public String statementImportPage(
-        @ModelAttribute("globalDateFilter") DateFilterState globalDateFilter,
-        Model model
-    ) {
-        fillStatementImportOptions(model, globalDateFilter);
-        model.addAttribute("mode", "idle");
-        return "entries/statement-import";
-    }
-
-    @PostMapping("/import/jobs")
-    public String startImport(
-        @RequestParam(name = "file", required = false) MultipartFile file,
-        @RequestParam(name = "headerOption", defaultValue = "WITH_HEADER") String headerOption,
-        @RequestParam(name = "separatorOption", defaultValue = "COMMA") String separatorOption,
-        @RequestParam(name = "separatorOther", required = false) String separatorOther,
-        @RequestParam(name = "dateFormatOption", defaultValue = "ISO") String dateFormatOption,
-        @RequestParam(name = "dateFormatOther", required = false) String dateFormatOther,
-        @RequestParam(name = "amountFormatOption", defaultValue = "PT_BR") String amountFormatOption,
-        @RequestParam(name = "amountFormatOther", required = false) String amountFormatOther,
-        Model model
-    ) {
-        try {
-            log.info("Starting entries CSV import. filename={}", file != null ? file.getOriginalFilename() : null);
-            EntryCsvImportOptions options = buildImportOptions(
-                headerOption,
-                separatorOption,
-                separatorOther,
-                dateFormatOption,
-                dateFormatOther,
-                amountFormatOption,
-                amountFormatOther
-            );
-            String jobId = entryImportJobCoordinator.startCsvJob(file, options);
-            fillImportJobModel(model, entryImportJobCoordinator.getSnapshot(jobId));
-        } catch (IllegalArgumentException ex) {
-            fillImportErrorModel(model, ex.getMessage());
-        }
-        return "entries/import :: result";
-    }
-
-    @PostMapping("/statement-import/jobs")
-    public String startStatementImport(
-        @RequestParam(name = "file", required = false) MultipartFile file,
-        @RequestParam(name = "accountId", required = false) Long accountId,
-        @RequestParam(name = "formatId", required = false) String formatId,
-        @ModelAttribute("globalDateFilter") DateFilterState globalDateFilter,
-        Model model
-    ) {
-        fillStatementImportOptions(model, globalDateFilter);
-        try {
-            String jobId = entryImportJobCoordinator.startStatementJob(file, accountId, formatId);
-            fillImportJobModel(model, entryImportJobCoordinator.getSnapshot(jobId));
-        } catch (IllegalArgumentException ex) {
-            fillImportErrorModel(model, ex.getMessage());
-        }
-        return "entries/statement-import :: result";
-    }
-
-    @GetMapping("/import/jobs/{jobId}")
-    public String importStatus(@PathVariable String jobId, HttpServletRequest request, Model model) {
-        EntryImportJobSnapshot snapshot = entryImportJobCoordinator.getSnapshot(jobId);
-        if (snapshot == null) {
-            String correlationId = String.valueOf(request.getAttribute(CorrelationIdFilter.CORRELATION_ID_KEY));
-            log.warn("Import job snapshot not found. correlationId={}, jobId={}", correlationId, jobId);
-            model.addAttribute("mode", "failed");
-            model.addAttribute("failureCauseKey", "entries.import.result.failure.unknown");
-            model.addAttribute("failedRow", null);
-            return "entries/import :: result";
-        }
-
-        fillImportJobModel(model, snapshot);
-        return "entries/import :: result";
-    }
-
-    @GetMapping("/statement-import/jobs/{jobId}")
-    public String statementImportStatus(
-        @PathVariable String jobId,
-        @ModelAttribute("globalDateFilter") DateFilterState globalDateFilter,
-        Model model
-    ) {
-        fillStatementImportOptions(model, globalDateFilter);
-        EntryImportJobSnapshot snapshot = entryImportJobCoordinator.getSnapshot(jobId);
-        if (snapshot == null) {
-            model.addAttribute("mode", "failed");
-            model.addAttribute("failureCauseKey", "entries.import.result.failure.unknown");
-            model.addAttribute("failedRow", null);
-            return "entries/statement-import :: result";
-        }
-
-        fillImportJobModel(model, snapshot);
-        return "entries/statement-import :: result";
     }
 
     @GetMapping("/new")
@@ -1002,13 +891,6 @@ public class EntryController {
         model.addAttribute("toastLevel", level);
     }
 
-    private void fillStatementImportOptions(Model model, DateFilterState globalDateFilter) {
-        DateFilterState effectiveDateFilter = resolveGlobalDateFilter(model, globalDateFilter);
-        model.addAttribute("statementAccountOptions", accountService.listVisibleForEntryFilter(effectiveDateFilter.getStartDate()));
-        List<EntryStatementFormat> formats = entryStatementImportService.listAvailableFormats();
-        model.addAttribute("statementFormats", formats);
-    }
-
     private void fillEntryFormAccountOptions(Model model, Long selectedAccountId) {
         List<Account> accounts = accountService.listAvailableForEntryForm(selectedAccountId);
         model.addAttribute("accountOptions", accounts);
@@ -1206,195 +1088,4 @@ public class EntryController {
         return transferViewsByEntryId;
     }
 
-    private void fillImportErrorModel(Model model, String rawMessage) {
-        model.addAttribute("mode", "failed");
-        model.addAttribute("failedRow", null);
-        model.addAttribute("failedColumnKey", null);
-        model.addAttribute("failureDetailMessage", rawMessage);
-        model.addAttribute("failureCauseKey", toFailureMessageKey(rawMessage, EntryImportFailureCause.UNKNOWN_ERROR));
-    }
-
-    private void fillImportJobModel(Model model, EntryImportJobSnapshot snapshot) {
-        if (snapshot == null) {
-            model.addAttribute("mode", "failed");
-            model.addAttribute("failureCauseKey", "entries.import.result.failure.unknown");
-            model.addAttribute("failedRow", null);
-            model.addAttribute("failedColumnKey", null);
-            model.addAttribute("failureDetailMessage", "Unknown import job");
-            return;
-        }
-
-        if (snapshot.status() == EntryImportJobStatus.PROCESSING) {
-            model.addAttribute("mode", "processing");
-            model.addAttribute("jobId", snapshot.jobId());
-            model.addAttribute("processedRows", snapshot.processedRows());
-            model.addAttribute("totalRows", snapshot.totalRows());
-            model.addAttribute("progressPercent", toProgressPercent(snapshot.processedRows(), snapshot.totalRows(), false));
-            return;
-        }
-
-        if (snapshot.status() == EntryImportJobStatus.SUCCESS) {
-            model.addAttribute("mode", "success");
-            model.addAttribute("summary", snapshot.summary());
-            model.addAttribute("progressPercent", 100);
-            return;
-        }
-
-        model.addAttribute("mode", "failed");
-        model.addAttribute("failedRow", snapshot.failedRow());
-        model.addAttribute("failedColumnKey", toColumnMessageKey(snapshot.failedColumn()));
-        model.addAttribute("failureDetailMessage", snapshot.failureMessage());
-        model.addAttribute("failureCauseKey", toFailureMessageKey(snapshot.failureMessage(), snapshot.failureCause()));
-        model.addAttribute("progressPercent", toProgressPercent(snapshot.processedRows(), snapshot.totalRows(), true));
-    }
-
-    private int toProgressPercent(int processedRows, int totalRows, boolean forceCompleteOnEmpty) {
-        if (totalRows <= 0) {
-            return forceCompleteOnEmpty ? 100 : 0;
-        }
-        return Math.min(100, (processedRows * 100) / totalRows);
-    }
-
-    private String toFailureMessageKey(String rawMessage, EntryImportFailureCause causeType) {
-        if (rawMessage != null) {
-            String normalizedMessage = rawMessage.toLowerCase();
-            if (normalizedMessage.contains("must not be empty")) {
-                return "entries.import.result.failure.emptyFile";
-            }
-            if (normalizedMessage.contains("only csv files are accepted")) {
-                return "entries.import.result.failure.invalidFileType";
-            }
-            if (normalizedMessage.contains("invalid separator")) {
-                return "entries.import.result.failure.invalidSeparator";
-            }
-            if (normalizedMessage.contains("invalid date format pattern")) {
-                return "entries.import.result.failure.invalidDateFormatPattern";
-            }
-            if (normalizedMessage.contains("invalid amount format pattern")) {
-                return "entries.import.result.failure.invalidAmountFormatPattern";
-            }
-            if (normalizedMessage.contains("account must be informed")) {
-                return "entries.import.result.failure.missingAccount";
-            }
-            if (normalizedMessage.contains("selected account must be active")) {
-                return "entries.import.result.failure.inactiveAccount";
-            }
-            if (normalizedMessage.contains("statement format must be informed")) {
-                return "entries.import.result.failure.missingStatementFormat";
-            }
-            if (normalizedMessage.contains("statement format is not supported")) {
-                return "entries.import.result.failure.invalidStatementFormat";
-            }
-        }
-
-        if (causeType == EntryImportFailureCause.MISSING_REQUIRED_FIELD) {
-            return "entries.import.result.failure.missingData";
-        }
-        if (causeType == EntryImportFailureCause.INVALID_FORMAT) {
-            return "entries.import.result.failure.invalidFormat";
-        }
-        return "entries.import.result.failure.unknown";
-    }
-
-    private String toColumnMessageKey(String columnName) {
-        if (!StringUtils.hasText(columnName)) {
-            return null;
-        }
-        return switch (columnName) {
-            case "account" -> "entries.import.column.account";
-            case "movementDate" -> "entries.import.column.movementDate";
-            case "settlementDate" -> "entries.import.column.settlementDate";
-            case "description" -> "entries.import.column.description";
-            case "category" -> "entries.import.column.category";
-            case "amount" -> "entries.import.column.amount";
-            case "notes" -> "entries.import.column.notes";
-            case "externalId" -> "entries.import.column.externalId";
-            default -> null;
-        };
-    }
-
-    private EntryCsvImportOptions buildImportOptions(
-        String headerOption,
-        String separatorOption,
-        String separatorOther,
-        String dateFormatOption,
-        String dateFormatOther,
-        String amountFormatOption,
-        String amountFormatOther
-    ) {
-        boolean hasHeader = resolveHasHeader(headerOption);
-        char separator = resolveSeparator(separatorOption, separatorOther);
-        String datePattern = resolveDatePattern(dateFormatOption, dateFormatOther);
-        String amountPattern = resolveAmountPattern(amountFormatOption, amountFormatOther);
-        return new EntryCsvImportOptions(separator, datePattern, amountPattern, hasHeader);
-    }
-
-    private boolean resolveHasHeader(String headerOption) {
-        return !"WITHOUT_HEADER".equalsIgnoreCase(headerOption);
-    }
-
-    private char resolveSeparator(String separatorOption, String separatorOther) {
-        if ("SEMICOLON".equalsIgnoreCase(separatorOption)) {
-            return ';';
-        }
-        if ("PIPE".equalsIgnoreCase(separatorOption)) {
-            return '|';
-        }
-        if ("TAB".equalsIgnoreCase(separatorOption)) {
-            return '\t';
-        }
-        if ("OTHER".equalsIgnoreCase(separatorOption)) {
-            return resolveCustomSeparator(separatorOther);
-        }
-        return ',';
-    }
-
-    private char resolveCustomSeparator(String rawValue) {
-        if (!StringUtils.hasText(rawValue)) {
-            throw new IllegalArgumentException("Invalid separator: blank");
-        }
-
-        String value = rawValue;
-        if ("\\t".equals(value.trim())) {
-            return '\t';
-        }
-
-        if (value.length() != 1) {
-            throw new IllegalArgumentException("Invalid separator: expected one character");
-        }
-
-        return value.charAt(0);
-    }
-
-    private String resolveDatePattern(String dateFormatOption, String dateFormatOther) {
-        if ("BR".equalsIgnoreCase(dateFormatOption)) {
-            return "dd/MM/uuuu";
-        }
-        if ("US".equalsIgnoreCase(dateFormatOption)) {
-            return "MM/dd/uuuu";
-        }
-        if ("DMY_DASH".equalsIgnoreCase(dateFormatOption)) {
-            return "dd-MM-uuuu";
-        }
-        if ("OTHER".equalsIgnoreCase(dateFormatOption)) {
-            if (!StringUtils.hasText(dateFormatOther)) {
-                throw new IllegalArgumentException("Invalid date format pattern");
-            }
-            return dateFormatOther.trim();
-        }
-        return "uuuu-MM-dd";
-    }
-
-    private String resolveAmountPattern(String amountFormatOption, String amountFormatOther) {
-        if ("US".equalsIgnoreCase(amountFormatOption)) {
-            return "^-?(?:\\d{1,3}(?:,\\d{3})+|\\d+)(?:\\.\\d{1,2})?$";
-        }
-        if ("OTHER".equalsIgnoreCase(amountFormatOption)) {
-            if (!StringUtils.hasText(amountFormatOther)) {
-                throw new IllegalArgumentException("Invalid amount format pattern");
-            }
-            return amountFormatOther.trim();
-        }
-        return "^-?(?:\\d{1,3}(?:\\.\\d{3})+|\\d+)(?:,\\d{1,2})?$";
-    }
 }
